Port to Win32 of recent changes.
git-svn-id: https://vdsf.svn.sourceforge.net/svnroot/vdsf/trunk@618 53c26a9f-4021-0410-9f73-c4284c400f72

+++ src/Engine/HashMap.c
@@ -655,7 +655,7 @@ int vdseHashMapInit( vdseHashMap         * pHashMap,
       return -1;
    }
    
-   pHashMap->numFields = pDefinition->numFields;
+   pHashMap->numFields = (uint16_t) pDefinition->numFields;
 
    ptr = (vdseFieldDef*) vdseMalloc( &pHashMap->memObject, 
                                      pHashMap->numFields* sizeof(vdseFieldDef),
